refactor: domain 모듈의 model 파일을 패키지로 분리
