+++ app/src/main/java/woowacourse/movie/datasource/ReservationDataSource.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.datasource
 
-import woowacourse.movie.domain.Reservation
 import woowacourse.movie.domain.dataSource.ReservationDataSource
+import woowacourse.movie.domain.model.Reservation
 
 class ReservationDataSource :
     ReservationDataSource {
