+++ domain/src/main/java/woowacourse/movie/domain/dataSource/MovieDataSource.kt
@@ -1,6 +1,6 @@
 package woowacourse.movie.domain.dataSource
 
-import woowacourse.movie.domain.Movie
+import woowacourse.movie.domain.model.Movie
 
 interface MovieDataSource {
     fun getData(): List<Movie>
