+++ domain/src/main/java/woowacourse/movie/domain/dataSource/ReservationDataSource.kt
@@ -1,6 +1,6 @@
 package woowacourse.movie.domain.dataSource
 
-import woowacourse.movie.domain.Reservation
+import woowacourse.movie.domain.model.Reservation
 
 interface ReservationDataSource {
     fun getData(): List<Reservation>
