+++ domain/src/main/java/woowacourse/movie/domain/discountPolicy/DiscountPolicy.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.domain.discountPolicy
 
-import woowacourse.movie.domain.Price
-import woowacourse.movie.domain.ReservationDetail
+import woowacourse.movie.domain.model.Price
+import woowacourse.movie.domain.model.ReservationDetail
 
 interface DiscountPolicy {
     fun discount(reservationDetail: ReservationDetail, price: Price): Price
