+++ domain/src/main/java/woowacourse/movie/domain/model/DateRange.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.model
 
 import java.time.LocalDate
