+++ domain/src/main/java/woowacourse/movie/domain/model/Image.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.model
 
 data class Image(
     val resource: Int
