+++ domain/src/main/java/woowacourse/movie/domain/model/ReservationDetail.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.model
 
 import java.time.LocalDateTime
