+++ domain/src/main/java/woowacourse/movie/domain/repository/MovieRepository.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.domain.repository
 
-import woowacourse.movie.domain.Movie
 import woowacourse.movie.domain.dataSource.MovieDataSource
+import woowacourse.movie.domain.model.Movie
 
 class MovieRepository(private val dataSource: MovieDataSource) {
     fun getData(): List<Movie> {
