+++ domain/src/main/java/woowacourse/movie/domain/repository/ReservationRepository.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.domain.repository
 
-import woowacourse.movie.domain.Reservation
 import woowacourse.movie.domain.dataSource.ReservationDataSource
+import woowacourse.movie.domain.model.Reservation
 
 class ReservationRepository(private val dataSource: ReservationDataSource) {
     fun getData(): List<Reservation> {
