+++ domain/src/main/java/woowacourse/movie/domain/seat/Seats.kt
@@ -1,10 +1,10 @@
 package woowacourse.movie.domain.seat
 
-import woowacourse.movie.domain.Price
-import woowacourse.movie.domain.ReservationDetail
 import woowacourse.movie.domain.discountPolicy.Discount
 import woowacourse.movie.domain.discountPolicy.MovieDay
 import woowacourse.movie.domain.discountPolicy.OffTime
+import woowacourse.movie.domain.model.Price
+import woowacourse.movie.domain.model.ReservationDetail
 
 data class Seats(val value: List<Seat>) {
     fun calculateDiscountedPrice(
